% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biva-package.R
\docType{package}
\name{biva-package}
\alias{BIVA}
\alias{BIVA-package}
\alias{biva-package}
\alias{biva}
\title{The 'biva' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Yichi Zhang \email{ychzhang@google.com}

}
