% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biva.R
\name{biva}
\alias{biva}
\title{Launch biva User Interface}
\usage{
biva()
}
\value{
Nothing
}
\description{
biva() loads interactive user interface built using R shiny.
}
\details{
Start biva

The interactive user interface is to provide an easy way for people who are learning business intelligence. Includes example data for testing out a few example analysis.
}
\examples{
\dontrun{
library(shiny)
biva()
}
}
\keyword{biva}
