% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biva.R
\docType{class}
\name{biva}
\alias{biva}
\title{Bayesian Instrument Variable Analysis}
\description{
Bayesian Instrument Variable Analysis

Bayesian Instrument Variable Analysis
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{version}}{BIVA package version used to fit model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-biva-new}{\code{biva$new()}}
\item \href{#method-biva-tracePlot}{\code{biva$tracePlot()}}
\item \href{#method-biva-weakIVTest}{\code{biva$weakIVTest()}}
\item \href{#method-biva-strataProbEstimate}{\code{biva$strataProbEstimate()}}
\item \href{#method-biva-calcProb}{\code{biva$calcProb()}}
\item \href{#method-biva-pointEstimate}{\code{biva$pointEstimate()}}
\item \href{#method-biva-credibleInterval}{\code{biva$credibleInterval()}}
\item \href{#method-biva-vizdraws}{\code{biva$vizdraws()}}
\item \href{#method-biva-lollipop}{\code{biva$lollipop()}}
\item \href{#method-biva-plotPrior}{\code{biva$plotPrior()}}
\item \href{#method-biva-predict}{\code{biva$predict()}}
\item \href{#method-biva-clone}{\code{biva$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-new"></a>}}
\if{latex}{\out{\hypertarget{method-biva-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$new(
  data,
  y,
  d,
  z,
  x_ymodel,
  x_smodel,
  y_type = "real",
  ER = 1,
  side = 2,
  beta_mean_ymodel,
  beta_sd_ymodel,
  beta_mean_smodel,
  beta_sd_smodel,
  sigma_shape_ymodel,
  sigma_scale_ymodel,
  seed = 1997,
  fit = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-tracePlot"></a>}}
\if{latex}{\out{\hypertarget{method-biva-tracePlot}{}}}
\subsection{Method \code{tracePlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$tracePlot(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-weakIVTest"></a>}}
\if{latex}{\out{\hypertarget{method-biva-weakIVTest}{}}}
\subsection{Method \code{weakIVTest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$weakIVTest()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-strataProbEstimate"></a>}}
\if{latex}{\out{\hypertarget{method-biva-strataProbEstimate}{}}}
\subsection{Method \code{strataProbEstimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$strataProbEstimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-calcProb"></a>}}
\if{latex}{\out{\hypertarget{method-biva-calcProb}{}}}
\subsection{Method \code{calcProb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$calcProb(a = 0, b = NULL, prior = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-pointEstimate"></a>}}
\if{latex}{\out{\hypertarget{method-biva-pointEstimate}{}}}
\subsection{Method \code{pointEstimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$pointEstimate(median = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-credibleInterval"></a>}}
\if{latex}{\out{\hypertarget{method-biva-credibleInterval}{}}}
\subsection{Method \code{credibleInterval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$credibleInterval(width = 0.75, round = 2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-vizdraws"></a>}}
\if{latex}{\out{\hypertarget{method-biva-vizdraws}{}}}
\subsection{Method \code{vizdraws()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$vizdraws(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-lollipop"></a>}}
\if{latex}{\out{\hypertarget{method-biva-lollipop}{}}}
\subsection{Method \code{lollipop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$lollipop(threshold = 0, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-plotPrior"></a>}}
\if{latex}{\out{\hypertarget{method-biva-plotPrior}{}}}
\subsection{Method \code{plotPrior()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$plotPrior()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-predict"></a>}}
\if{latex}{\out{\hypertarget{method-biva-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$predict(new_data, name = NULL, M = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biva-clone"></a>}}
\if{latex}{\out{\hypertarget{method-biva-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biva$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
