% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{CleanData}
\alias{CleanData}
\title{Cleans and prepares data for analysis}
\usage{
CleanData(
  data,
  y,
  d,
  z,
  x_ymodel = NULL,
  x_smodel = NULL,
  ER = ER,
  side = side
)
}
\arguments{
\item{data}{A data.frame containing the data to be cleaned.}

\item{y}{Name of the dependent variable (character; numeric).}

\item{d}{Name of the treatment received variable (character; numeric 0 or 1).}

\item{z}{Name of the treatment assigned variable (character; numeric 0 or 1).}

\item{x_ymodel}{Names of the covariates to include in the Y-model (character vector, optional).}

\item{x_smodel}{Names of the covariates to include in the S-model (character vector, optional).}

\item{ER}{The assumption of exclusion restriction (numeric; 1 if assumed, 0 otherwise).}

\item{side}{The number of noncompliance sides (numeric; 1 or 2).}
}
\value{
A list containing the cleaned dataset and relevant metadata:
\itemize{
\item \code{N}: The number of observations after cleaning.
\item \code{P_ymodel} The number of covariates in the Y-model after cleaning.
\item \code{P_smodel} The number of covariates in the S-model after cleaning.
\item \code{Z} The treatment assigned vector (instrument variable).
\item \code{D} The treatment received vector (treatment variable).
\item \code{Y} The dependent variable vector (outcome variable).
\item \code{X_ymodel} The cleaned covariate matrix in the Y-model (including intercept).
\item \code{X_smodel} The cleaned covariate matrix in the S-model (including intercept).
\item \code{K_ymodel} The number of Y-models.
\item \code{K_smodel} The number of S-models.
}
}
\description{
This function performs a series of data cleaning and preprocessing steps
to ensure the data is suitable for analysis.
}
